Move `look_for_tests` to `private_items_doc_tests`
